\name{lcy.cor}
\alias{lcy.cor}
\title{
    compute correlation
}
\description{
    compute correlation between two variables. the build-in cor function can not be applied to the variables containing missing values.
    it is the same if it uses "use" parameter in cor function..
}
\usage{
lcy.cor(x, y, ...)
}
\arguments{
  \item{x}{
    numeric vector
}
  \item{y}{
    numeric vector
}
  \item{\dots}{
    see cor for more details
}
}
\details{
    see cor for more details
}
\value{
    values corresponding to the function.
}
\author{
    Chengyu Liu <chengyu.liu@helsinki.fi>
}
\keyword{ cor }
\keyword{ statistics }
