\name{lcy.initiation}
\alias{lcy.initiation}
\title{
    lcy.initiation
}
\description{ used to create a new recursive directory to save results. If it exists, remove and create a new one.
}
\usage{
    lcy.initiation(path = "./")
}
\arguments{
  \item{path}{ directory to be created recursively. 
}
}
\author{
    Chengyu Liu <chengyu.liu@helsinki.fi>
}
\keyword{ initiation }
