\name{lcy.matrix.in}
\alias{lcy.matrix.in}
\title{
    lcy is matrix in
}
\description{
    check whether rows or columns in m1 exits in m2. return logic value showing whether the element of m1 exist in m2 or not. 
}
\usage{
lcy.is.matrix.in(m1, m2, order = TRUE, byrow = TRUE)
}
\arguments{
  \item{m1}{ a matrix
}
  \item{m2}{ a matrix
}
  \item{order}{ logic value showing whether order of the element in each row/column should be considered. TRUE for yes.
}
  \item{byrow}{ logic value showing whether comparison is done by row or column.
}
}
\value{
    a logic vector
}
\author{
    chengyu.liu@helsinki.fi
}
\keyword{ }
