\name{lcy.misc.cutdendrogram}
\alias{lcy.misc.cutdendrogram}
\title{
    cut dendrogram
}
\description{
    used to cut dendrigran and return the labels of each cluster
}
\usage{
lcy.misc.cutdendrogram(dendrogram, h)
}
\arguments{
  \item{dendrogram}{
    dendrogram object which is a component of the return of heatmap.
}
  \item{h}{
    a numeric value
}
}
\value{
    a list of cluster leaves.
}
\author{
    Chengyu Liu <chengyu.liu.cs@gmail.com>
}
\keyword{ cut }
\keyword{ dendrogram }
